+++ radtools/crystal/kpoints.py
@@ -392,277 +392,132 @@ class Kpoints:
                     flatten_points = np.concatenate((flatten_points, delta))
         return flatten_points
 
-#function checking if the k points in the list k_list are inside the Brillouin zone
+###function checking if the k points in the list k list are inside the BZ
 def downfold_inside_brillouin_zone(
-    k_point_list,
-    brillouin_primitive_vectors_3d):
-    r"""
-        check if any point in a list is inside the first brillouin zone, in case the point is outside it is shifted back to the first brillouin zone
-        Parameters
-        ----------
-        k_point_list: (N, 3) :|array_like| kx,ky,kz (k points are given in cartesian coordinates)
-        brillouin_primitive_vectors_3d: (3,3) :|array_like| columns: kx,ky,kz, rows: b1,b2,b3
-        Returns
-        -------
-        transformed_k_point_list: (N, 3) :|array_like| kx,ky,kz (k points are given in cartesian coordinates)
-        """
-    matrix_transformation_crystal_to_cartesian=np.zeros((3,3))
-    matrix_transformation_cartesian_to_crystal=np.zeros((3,3))
-    number_k_point_elements = k_point_list.shape[0]
-    transformed_k_point_list = np.zeros((number_k_point_elements,3),dtype=float)
-    matrix_transformation_crystal_to_cartesian=np.matrix(brillouin_primitive_vectors_3d).transpose()
+    k_points_list,
+    brillouin_primitive_vectors_3D):
+   
+    matrix_transformation_crystal_to_cartesian=np.zeros((3,3),dtype=float)
+    matrix_transformation_cartesian_to_crystal=np.zeros((3,3),dtype=float)
+    
+    number_k_points = k_points_list.shape[0]
+    transformed_k_points_list = np.zeros((number_k_points,3),dtype=float)
+    
+    matrix_transformation_crystal_to_cartesian=np.matrix(brillouin_primitive_vectors_3D).transpose()
     matrix_transformation_cartesian_to_crystal=np.linalg.inv(matrix_transformation_crystal_to_cartesian)
 
-    # writing k points in crystal coordinates
-    for i in range(number_k_point_elements):
-        transformed_k_point_list[i,:]=matrix_transformation_cartesian_to_crystal@k_point_list[i,:]
+    ### writing k points in crystal coordinates
+    for i in range(number_k_points):
+        transformed_k_points_list[i,:]=matrix_transformation_cartesian_to_crystal@k_points_list[i,:]
 
-    # checking if any point of the list is outside the first brillouin zone
-    #to each k point is associated a int 0,1,2,3 if it goes over the bonds
+    ### checking if any point of the list is outside the first brillouin zone
+    ### to each k point is associated a int 0,1,2,3 if it goes over the boundaries
     indices=[]
-    for i in range(number_k_point_elements):
+    for i in range(number_k_points):
         count=0
         for r in range(3):
-            if transformed_k_point_list[i,r] >= 1.0 or transformed_k_point_list[i,r] < 0:
+            if transformed_k_points_list[i,r]>=1.0 or transformed_k_points_list[i,r]<0:
                 count+=r
                 for d in range(3):
-                    k_point_list[i,d]=k_point_list[i,d]-brillouin_primitive_vectors_3d[r,d]*np.sign(transformed_k_point_list[i,r])
+                    k_points_list[i,d]=k_points_list[i,d]-brillouin_primitive_vectors_3D[r,d]*np.sign(transformed_k_points_list[i,r])
         indices.append(count)
-    indices=np.reshape(indices,(number_k_point_elements,1))
-    return k_point_list,indices
+    return k_points_list,indices
 
-# function applying symmetry analysis to a list of k points, the symmetry operations considered are the point group ones
-# the fixed point is given; the analysis aim is to redistribute the fixed point weight between the list of k points
-# the found symmetry orbits are counted in the redistribution of the fixed point weight a number of times equal to the number of k points inside the orbit itself
+### function applying symmetry analysis to a list of k points, the symmetry operations considered are the point group ones
+### the fixed point is given; the analysis aim is to redistribute the fixed point weight between the list of k points
+### the number of found symmetry orbits are counted in the redistribution of the fixed point weight
 def symmetry_analysis(
-    k_origin,
-    k_origin_weight,
-    k_list,
+    k_fixed_point,
+    k_fixed_point_weight,
+    k_points_list,
     symmetries,
-    threshold,
+    threshold_symmetry
 ):
-    r"""
-    Given a list of k points "k_list" and an origin "k_origin" with a certain weight "k_origin_weight", different symmetry operations are applied to the list of k points, keeping the origin fixed.
-    The origin weight is propelry distributed between the points:
-        the orbits of the symmetry operations acquire a certain percentage of the origin weight, which takes into account how many orbits are found 
-        of the orbits one k point (from the k list) is chosen as representative, and the weight of the orbit is assigned only to it
-    Parameters
-    ----------
-    k_origin:(3) |array_like| fixed point coordinates (kx,ky,kz), fixed point considered in the point-group symmetry operations
-    k_origin_weight : |float|  weight to distribute between the k points of the list 
-    k_list : (N, 3) |array_like| list of k points considered in the symmetry analysis (kx,ky,kz)
-    symmetries : list of lists, a symmetry is a list of 3 elements (the versor is the axis of rotation, while the modulus is the angle).
-    threshold : |float| threshold to recognize a symmetry.
-    Returns
-    -------
-    new_k_list: (N,4) :|array_like| list of k points coordinates with respcetive weights from the symmetry analysis
-    """
-    number_elements = k_list.shape[0]  
+    number_elements = k_points_list.shape[0]  
 
-    new_k_list = np.c_[ k_list, (k_origin_weight / number_elements) * np.ones(number_elements, dtype=float)]
-    # if there are no symmetry operations, than the weight on each k point is exactly 1/N of the original weight
-    if symmetries is None or (symmetries[0][0] == symmetries[0][1] == symmetries[0][2] == 0):
-        return new_k_list
-
-    # defining a matrix to save the degeneracies, after each symmetry operation
-    check_degeneracy = np.zeros((number_elements, number_elements), dtype=bool)
-    # saving flag if no degeneracy is detected
-    no_degeneracy = True
+    new_k_points_list = np.c_[ k_points_list, (k_fixed_point_weight / number_elements) * np.ones(number_elements, dtype=float)]
     
-    k_eigenspaces=np.zeros(number_elements)
-    for i in range(number_elements):
-        k_eigenspaces[i]=i
+    ### if there are no symmetry operations, the weight on each k point is exactly 1/(number_elements) of the original weight
+    if ((symmetries is None) or (symmetries[0][0] == symmetries[0][1] == symmetries[0][2] == 0)):
+        return new_k_points_list
 
-    # each k point is compared to each other k point after each of the symmetry operations has been applied
-    for i in range(number_elements - 1):
-        for j in range(i + 1, number_elements):
-            # for each pair considering all the symmetry operations
+    ### defining a matrix to save the degeneracies, after each symmetry operation
+    check_degeneracy = np.zeros((number_elements, number_elements), dtype=bool)
+    ### saving flag if no degeneracy is detected
+    no_degeneracies = True
+    
+    eigenspaces=np.zeros(number_elements,dtype=int)
+    for i in range(number_elements):
+        eigenspaces[i]=i
+    ### each k point is compared to each other k point after each of the symmetry operations has been applied
+    for i in range(number_elements-1):
+        for j in range(i+1,number_elements):
+            ### for each pair considering all the symmetry operations
             for rotvec in symmetries:
-                k_point_transformed = (
+                transformed_k_point = (
                     Rotation.from_rotvec(rotvec).as_matrix()
-                    @ (k_list[j] - k_origin)
-                    + k_origin
+                    @ (k_points_list[j] - k_fixed_point)
+                    + k_fixed_point
                 )
-                # if any degeneracy is detected, the check on all the symmetry operations given is stopped
+                ### if any degeneracy is detected, the check on all the symmetry operations given is stopped
                 if np.allclose(
-                    k_list[i], k_point_transformed, atol=threshold
+                    k_points_list[i], transformed_k_point, atol=threshold_symmetry
                 ):
                     check_degeneracy[i][j] = True
-                    no_degeneracy = False
+                    no_degeneracies = False
                     break
 
-    # if no degeneracy is detected, the no-symmetry-operations result is given
-    if no_degeneracy:
-        return new_k_list
+    ### if no degeneracy is detected, the same result as in the case of no symmetry operation is given
+    if no_degeneracies:
+        return new_k_points_list
     else:
-        # if degeneracy is detected, of the degenerate points only one is chosen as representative of the respective orbit (the first to appear in the k list)
-        # assuming all k points as not-degenerate, and creating a dictionary
-        # where to each eigenspace (orbit) is associated a not null k_point
-        for i in range(0,number_elements-1):
+        ### if degeneracy is detected, between the degenerate points only one is chosen as representative of the respective orbit (the first to appear in the k points list)
+        ### at first assuming all k points as not-degenerate
+        ### to each eigenspace (orbit) is associated a not null k_point
+        for i in range(number_elements-1):
             for j in range(i+1,number_elements):
                 if check_degeneracy[i][j]==True:
                     min_value=min(i,j)
                     max_value=max(i,j)
-                    k_eigenspaces[k_eigenspaces==max_value]=min_value
+                    eigenspaces[eigenspaces==max_value]=min_value
 
-        # counting number of different eigenspaces
-        number_eigenspaces=len(np.unique(k_eigenspaces))
-        weights=k_origin_weight/number_eigenspaces
+        ### counting number of different eigenspaces
+        number_eigenspaces=len(np.unique(eigenspaces))
+        weights=k_fixed_point_weight/number_eigenspaces
 
-        # ordering eigenspaces values
-        ordered_k_eigenspaces=np.zeros(number_elements)
+        ### ordering eigenspaces values
+        ordered_eigenspaces=np.zeros(number_elements)
         counting=0
         assigned_indices=[]
         for i in range(number_elements):
-            if k_eigenspaces[i] not in assigned_indices:
+            if eigenspaces[i] not in assigned_indices:
                 if counting<number_eigenspaces:
-                    ordered_k_eigenspaces[k_eigenspaces==k_eigenspaces[i]]=counting
-                    assigned_indices.extend([r for r in range(number_elements) if k_eigenspaces[r]==k_eigenspaces[i]])
+                    ordered_eigenspaces[eigenspaces==eigenspaces[i]]=counting
+                    assigned_indices.extend([r for r in range(number_elements) if eigenspaces[r]==eigenspaces[i]])
                     counting+=1
-        # considering a representative for each eigenspace
+        ### considering a representative for each eigenspace
         for i in range(number_eigenspaces):
-            list_positions=[r for r in range(number_elements) if ordered_k_eigenspaces[r]==i]
+            list_positions=[r for r in range(number_elements) if ordered_eigenspaces[r]==i]
             if len(list_positions)!=0:
-                new_k_list[list_positions[0],3]=weights
-                new_k_list[list_positions[1:],3]=0
+                new_k_points_list[list_positions[0],3]=weights
+                new_k_points_list[list_positions[1:],3]=0
             else:
                 break
 
-        # returning the matrix with the respective weights
-        return new_k_list
+        ### returning the k points and their respective weights
+        return new_k_points_list
 
-# local refinment of the k list in the plane (parallelepiped) pointed out by the brillouin primitive vectors given (it can be 2D, 3D or ..)
-def local_refinment_with_symmetry_analysis(
-    new_k_list,
-    old_k_list,
-    refinment_spacing,
-    refinment_iteration,
-    symmetries,
-    threshold,
-    brillouin_primitive_vectors_3d,
-    brillouin_primitive_vectors_2d,
-    normalized_brillouin_primitive_vectors_2d,
-    downfold
-):
-    r"""
-    Starting from a set of k points (old_k_list) with certain weights, to each k point iteratively (refinment_iteration) a new subset of k points is associated,
-    a symmetry analysis (symmetry) is performed on the subset, and consequently the initial weight of the k point is distributed in the subset of k points.
-    This procedure goes on for the given number of refinment iterations.
-    At each iteration for each k point four new k points are generated along the reciprocal primitive vectors of the selcted plane (2d refinment), at a distance from the original k point 
-    equal to the refinment spacing (the refinment spacing is halved at each iteration).
-    Obviously no refinment is applied to k points with null weight.
-
-    if check_inside_dominion == True instead of checking that the point is inside the 2d plane of the brillouin zone, it is checked that it is inside a circle of a certain radius
-    (radius), and if it is outside the circle it is downfolded back 
-
-    Parameters
-    ----------
-    new_k_list: (,4) |array_like| list of values produced by the refinment procedure (expected to be empty at the beginning of the refinment procedure)
-    old_k_list: (,4) |list| list of values give to the refinment procedure (kx,ky,kz,weight)
-    refinment_spacing: |double| initial spacing given to generate the refinment, half of the preceding refinment spacing is considered at each refinment iteration
-    refinment_iteration: |int| number of refinment iterations considered
-    symmetry: |list of lists| a symmetry is a list of 3 elements (the versor is the axis of rotation, while the modulus is the angle)
-    threshold: |double| threshold to recognize a symmetry
-    brillouin_primitive_vectors : (2,3) |array_like| vectors definining the 2D plane in the reciprocal space, chosen for the refinment procedure
-    brillouin_primitive_vectors_all: (3,3) |array_like| vectors definining the 3D reciprocal space
-    (the k points of the list are expected to be in the same 2D plane)
-    """
-    length_old_k_list=old_k_list.shape[0]
-    if refinment_iteration == 0:
-        if downfold == True:
-            old_k_list[:,:3]= downfold_inside_brillouin_zone(
-                old_k_list[:,:3],
-                brillouin_primitive_vectors_3d,
-            )
-        for i in range(length_old_k_list):
-            if old_k_list[i,3]!=0:
-                new_k_list.extend(old_k_list[i,:])
-    else:
-        iter = 0
-        while iter != length_old_k_list:
-            # reading the k points inside the list and refine around them
-            if length_old_k_list == 1:
-                k_tmp = old_k_list[0,:3]
-                k_tmp_weight = old_k_list[0,3]
-                iter = length_old_k_list
-            else:
-                k_tmp = old_k_list[iter,:3]
-                k_tmp_weight = old_k_list[iter,3]
-                iter = iter + 1
-            # refinment procedure is applied to the single k point if its weight is not null
-            if k_tmp_weight != 0:
-                # a subgrid of points is associated to each k point
-                k_tmp_subgrid = np.zeros((4, 4))
-                k_tmp_subgrid[:,:3] = k_tmp
-                k_tmp_subgrid[0,:3] +=  normalized_brillouin_primitive_vectors_2d[0,:]*refinment_spacing
-                k_tmp_subgrid[1,:3] -=  normalized_brillouin_primitive_vectors_2d[0,:]*refinment_spacing
-                k_tmp_subgrid[2,:3] +=  normalized_brillouin_primitive_vectors_2d[1,:]*refinment_spacing
-                k_tmp_subgrid[3,:3] -=  normalized_brillouin_primitive_vectors_2d[1,:]*refinment_spacing
-                
-                k_tmp_subgrid = symmetry_analysis(
-                    k_tmp,
-                    k_tmp_weight,
-                    k_tmp_subgrid[:,:3],
-                    symmetries,
-                    threshold
-                )
-                
-                k_tmp_subgrid=np.delete(k_tmp_subgrid, np.where(k_tmp_subgrid[:,3]==0),axis=0)
-                if downfold == True:
-                    k_tmp_subgrid[:,:3]=downfold_inside_brillouin_zone(
-                        k_tmp_subgrid[:,:3],
-                        brillouin_primitive_vectors_3d,
-                    )
-                # applying to the points of the subgrid to the refinment procedure
-                new_refinment_spacing = refinment_spacing / 2
-                
-                local_refinment_with_symmetry_analysis(
-                    new_k_list,
-                    k_tmp_subgrid,
-                    new_refinment_spacing,
-                    refinment_iteration-1,
-                    symmetries,
-                    threshold,
-                    brillouin_primitive_vectors_3d,
-                    brillouin_primitive_vectors_2d,
-                    normalized_brillouin_primitive_vectors_2d,
-                    downfold
-                )
-
-def k_points_triangulation_2d(
+def k_points_triangulation_2D(
     k_points_list,
-    brillouin_primitive_vectors_2d,
-    count
+    brillouin_primitive_vectors_2D,
+    precedent_count,
+    border_indices=[]
 ):
-    r"""
-    the sparse k points (kx,ky,kz,w) are linked togheter through a triangulation procedure (Delaunay procedure)
-    the so-built triangles are listed, each triangle is characterized by three indices pointing to the vertices position in the k poins list  
-    the ordering of the vertices is so to assure a cloack-wise path along the triangles
-
-    the periodicity of the BZ is properly taken into account:
-        the points at the left border or bottom border are considered twice in the triangulation
-        once on their proper border, then on the opposite border (translation of a reciprocal vector)
-        the index of the triangulation is the same both times, this assures the periodicity condition
-    
-    Parameters
-    ---------
-    k_points_list: (n,4) |array_like| (kx,ky,kz,w)
-    brillouin_primitive_vectors_2d: (2,3) |array_like|
-    count: (int) taking into account if the k points need to be added to a list, so numbering properly the vertices, in order to be able to use these triangles with the new list
-    border_count: (int) taking into account if the first k points in k_points_list are from the border and are not neeeded to be added to the list  
-
-    Returns
-    ---------
-    k_points_list: (n,4) |array_like| (kx,ky,kz,w)
-    triangles: (s,3) |array_like| 
-    left_border_points, bottom_border_points indices of the points at the border of the BZ
-    """
-
     number_elements=k_points_list.shape[0]
     k_points_list_projections=np.zeros((number_elements,2),dtype=float)
-    #choosing one point as an origin in the 2d brillouin plane
+    ###choosing one point as an origin in the 2D brillouin plane
     origin=np.zeros(3,dtype=float)
-    #calculating the projections of the different k points on the primitive vectors
+    ###calculating the projections of the different k points on the primitive vectors
     vector_0=np.zeros((number_elements,2),dtype=float)
     vector_1=np.zeros((number_elements,2),dtype=float)
     vector_2=np.ones((number_elements,2),dtype=float)
@@ -670,285 +525,314 @@ def k_points_triangulation_2d(
         vector_0[i,0]=1
         vector_1[i,1]=1
         for j in range(2):
-            k_points_list_projections[i,j]=(k_points_list[i,:3]-origin)@brillouin_primitive_vectors_2d[j,:]
+            k_points_list_projections[i,j]=(k_points_list[i,:3]-origin)@brillouin_primitive_vectors_2D[j,:]
+    ###considering the periodic images as well
+    k_points_list_projections_with_periodic_images=np.zeros((4*number_elements,2),dtype=float)
+    k_points_list_projections_with_periodic_images[:number_elements,:]=k_points_list_projections
+    k_points_list_projections_with_periodic_images[number_elements:2*number_elements,:]=k_points_list_projections+vector_0
+    k_points_list_projections_with_periodic_images[2*number_elements:3*number_elements,:]=k_points_list_projections+vector_1
+    k_points_list_projections_with_periodic_images[3*number_elements:,:]=k_points_list_projections+vector_2
     
-    #considering the periodic images
-    k_points_list_projections_periodic_images=np.zeros((4*number_elements,2),dtype=float)
-    k_points_list_projections_periodic_images[:number_elements,:]=k_points_list_projections
-    k_points_list_projections_periodic_images[number_elements:2*number_elements,:]=k_points_list_projections+vector_0
-    k_points_list_projections_periodic_images[2*number_elements:3*number_elements,:]=k_points_list_projections+vector_1
-    k_points_list_projections_periodic_images[3*number_elements:,:]=k_points_list_projections+vector_2
-    
-    #triangulation of the set of points
-    old_triangles_tmp = Delaunay(k_points_list_projections_periodic_images,qhull_options="QJ")
-    print(old_triangles_tmp.add_points)
-
-    old_triangles_tmp=old_triangles_tmp.simplices
-    number_triangles = int(len(old_triangles_tmp))
-    old_triangles=[]
+    ###triangulation of the set of points
+    triangles_tmp = Delaunay(k_points_list_projections_with_periodic_images,qhull_options="QJ")
+    triangles_tmp=triangles_tmp.simplices
+    ###each vertex has a pair of elements: the nuber of the k point associated, and if it the traingle is there passing through the BZ border
+    number_triangles = int(len(triangles_tmp))
+    triangles=[]
     for i in range(number_triangles):
-        triangle=np.zeros((3,2),dtype=int)
+        single_triangle=np.zeros((3,2),dtype=int)
         count_element=0
         count_outside=0
-        for element in old_triangles_tmp[i]:
+        for element in triangles_tmp[i]:
             if element<number_elements:
-                triangle[count_element,1]=0
-                triangle[count_element,0]=element
+                single_triangle[count_element,1]=0
+                single_triangle[count_element,0]=element
             elif element>=number_elements and element<2*number_elements:
-                triangle[count_element,1]=1
+                single_triangle[count_element,1]=1
+                single_triangle[count_element,0]=element-number_elements
                 count_outside+=1
-                triangle[count_element,0]=element-number_elements
             elif element>=2*number_elements and element<3*number_elements:
-                triangle[count_element,1]=2
+                single_triangle[count_element,1]=2
+                single_triangle[count_element,0]=element-2*number_elements
                 count_outside+=1
-                triangle[count_element,0]=element-2*number_elements
             else:
-                triangle[count_element,1]=3
+                single_triangle[count_element,1]=3
+                single_triangle[count_element,0]=element-3*number_elements
                 count_outside+=1
-                triangle[count_element,0]=element-3*number_elements
             count_element+=1
         if count_outside < 3:
-            old_triangles.append(triangle)
-    number_triangles = int(len(old_triangles))
+            triangles.append(single_triangle)
+    ###counting the proper number of triangles
+    number_triangles = int(len(triangles))
 
-    #the projections along the two axis can be used to properly order the vertices of the triangls
-    new_triangles = []
+    ###TESTING NEEDED HERE
+    ###the projections along the two axis can be used to properly order the vertices of the triangls
+    ordered_triangles = []
     vertices_projections = np.zeros((3,2),dtype=float)
     one=[1,0]
     two=[0,1]
     three=[1,1]
     for i in range(number_triangles):
-        ##print(i,number_triangles)
-        ##print(old_triangles[i])
-        vertices=old_triangles[i]
-        ##for r in range(3):
-        ##    if vertices[r,1]==0:
-        ##        vertices_projections[r]=k_points_list_projections[vertices[r,0]]
-        ##    elif vertices[r,1]==1:
-        ##        vertices_projections[r]=k_points_list_projections[vertices[r,0]]+one
-        ##    elif vertices[r,1]==2:
-        ##        vertices_projections[r]=k_points_list_projections[vertices[r,0]]+two
-        ##    else:
-        ##        vertices_projections[r]=k_points_list_projections[vertices[r,0]]+three
-        ##indices=[tup[0] for tup in sorted(enumerate(vertices_projections[:,1]))]
-        ##vertices_projections=vertices_projections[indices]
-        ##if vertices_projections[0,1]==np.min(vertices_projections[:,0]):
-        ##    indices=[tup[0] for tup in sorted(enumerate(vertices_projections[:,0]))]
-        ##    vertices_projections=vertices_projections[indices]
-        ##vertices=vertices[indices]
-        #considering the case the function is called on an existing list (the existing list has length equal to count)
-        ##for s in range(3):
-        ##    if  vertices[s,0]> count:
-        ##        vertices[s,0]+=count
-        new_triangles.append(vertices)
-       
-    #for each triangle the ordering of the vertices is now clock-wise
-    return k_points_list,new_triangles
+        vertices=triangles[i]
+        for r in range(3):
+            if vertices[r,1]==0:
+                vertices_projections[r]=k_points_list_projections[vertices[r,0]]
+            elif vertices[r,1]==1:
+                vertices_projections[r]=k_points_list_projections[vertices[r,0]]+one
+            elif vertices[r,1]==2:
+                vertices_projections[r]=k_points_list_projections[vertices[r,0]]+two
+            else:
+                vertices_projections[r]=k_points_list_projections[vertices[r,0]]+three
+        indices=[tup[0] for tup in sorted(enumerate(vertices_projections[:,1]))]
+        vertices_projections=vertices_projections[indices]
+        if vertices_projections[0,1]==np.min(vertices_projections[:,0]):
+            indices=[tup[0] for tup in sorted(enumerate(vertices_projections[:,0]))]
+            vertices_projections=vertices_projections[indices]
+        vertices=vertices[indices]
+        ###considering the case the function is called on an existing list (the existing list has length equal to precedent_count)
+        border_count=number_elements-len(border_indices)
+        list_tmp=[]
+        for s in range(3):
+            if  vertices[s,0]> precedent_count and vertices[s,0]<border_count:
+                vertices[s,0]+=precedent_count
+        ###considering the case in which the function is called on new k points plus old k points, and the proper indices of the old k points are given by border indices
+        ###assuming as well that the old k points are after the new k points (so given the total number of k points the old ones are the ones > all_k_points-len(border_indices))
+            if vertices[s,0]>=border_count and border_count!=number_elements:
+                vertices[s,0]=border_indices[vertices[s,0]-border_count]
+            list_tmp.append(list(vertices[s]))
+        
+        ordered_triangles.append(list_tmp)
+    
+    ordered_triangles=np.reshape(ordered_triangles,(number_triangles,3,2))
+    ###for each triangle the ordering of the vertices is now clock-wise
+    #return k_points_list,ordered_triangles
+    return k_points_list,ordered_triangles
+
+### local refinment of the k points list in the 2D plane
+def local_refinment_with_symmetry_analysis(
+    old_k_points_list,
+    new_k_points_list,
+    refinment_spacing,
+    refinment_iterations,
+    symmetries,
+    threshold_symmetry,
+    brillouin_primitive_vectors_3d,
+    brillouin_primitive_vectors_2d,
+    normalized_brillouin_primitive_vectors_2d,
+    downfold_procedure=False
+):
+    length_old_k_points_list=old_k_points_list.shape[0]
+    if  refinment_iterations == 0:
+        if downfold_procedure == True:
+            old_k_points_list[:,:3],indices= downfold_inside_brillouin_zone(
+                old_k_points_list[:,:3],
+                brillouin_primitive_vectors_3d
+            )
+        for i in range(length_old_k_points_list):
+            if old_k_points_list[i,3]!=0.0:
+                new_k_points_list.extend(old_k_points_list[i,:])
+    else:
+        iter = 0
+        while iter != length_old_k_points_list:
+            ### reading the k points inside the list and refine around them
+            if length_old_k_points_list == 1:
+                k_point_tmp = old_k_points_list[0,:3]
+                k_point_tmp_weight = old_k_points_list[0,3]
+                iter = length_old_k_points_list
+            else:
+                k_point_tmp = old_k_points_list[iter,:3]
+                k_point_tmp_weight = old_k_points_list[iter,3]
+                iter = iter + 1
+            ### refinment procedure is applied to the single k point if its weight is not zero
+            if k_point_tmp_weight != 0:
+                ### a sub_grid of points is associated to each k point
+                k_point_tmp_subgrid = np.zeros((4, 4),dtype=float)
+                k_point_tmp_subgrid[:,:3] = k_point_tmp
+                k_point_tmp_subgrid[0,:3] +=  normalized_brillouin_primitive_vectors_2d[0,:]*refinment_spacing
+                k_point_tmp_subgrid[1,:3] -=  normalized_brillouin_primitive_vectors_2d[0,:]*refinment_spacing
+                k_point_tmp_subgrid[2,:3] +=  normalized_brillouin_primitive_vectors_2d[1,:]*refinment_spacing
+                k_point_tmp_subgrid[3,:3] -=  normalized_brillouin_primitive_vectors_2d[1,:]*refinment_spacing
+                ### the weight of the k point is redistributed on the subgrid taking into account a symmetry analysis
+                k_point_tmp_subgrid = symmetry_analysis(
+                    k_point_tmp,
+                    k_point_tmp_weight,
+                    k_point_tmp_subgrid[:,:3],
+                    symmetries,
+                    threshold_symmetry
+                )
+                ### deleting the k points of the subgrid having weight equal to zero
+                k_point_tmp_subgrid=np.delete(k_point_tmp_subgrid, np.where(k_point_tmp_subgrid[:,3]==0),axis=0)
+                
+                ### applying downfolding procedure
+                if downfold_procedure == True:
+                    k_point_tmp_subgrid[:,:3],indices=downfold_inside_brillouin_zone(
+                        k_point_tmp_subgrid[:,:3],
+                        brillouin_primitive_vectors_3d,
+                    )
+
+                ### applying reiteratively the refinment procedure
+                new_refinment_spacing = refinment_spacing / 2
+
+                local_refinment_with_symmetry_analysis(
+                    k_point_tmp_subgrid,
+                    new_k_points_list,
+                    new_refinment_spacing,
+                    refinment_iterations-1,
+                    symmetries,
+                    threshold_symmetry,
+                    brillouin_primitive_vectors_3d,
+                    brillouin_primitive_vectors_2d,
+                    normalized_brillouin_primitive_vectors_2d,
+                    downfold_procedure
+                )
 
 #generation of a 2D k points grid 
 def k_points_generator_2D(
-    brillouin_primitive_vectors_3d,
-    other_brillouin_primitive_vectors_3d,
-    initial_grid_spacing,
+    new_brillouin_primitive_vectors_3d,
+    old_brillouin_primitive_vectors_3d,
     chosen_plane,
-    brillouin_primitive_vectors_2d,
-    default_gridding=10,
-    symmetry_analysis_flag=False,
+    grid_spacing,
+    covering_BZ=False,
+    default_gridding=100,
+    symmetry_analysis=False,
     refinment_spacing=0,
+    refinment_iterations=0,
     symmetries=[[0,0,0]],
-    threshold=0.0001,
-    refinment_iteration=0,
+    threshold_symmetry=0.0001,
     shift_in_plane=[0,0],
     shift_in_space=[0,0,0],
     precedent_count=0
 ):
-    r"""
-    k points generator in a 2d plane
-    
-    Parameters
-    ----------
-    brillouin_primitive_vectors_3d: (3,3) |array_like| (columns: kx,ky,kz, rows: b1,b2,b3)
-    chosen_plane: (,3) |array_like|  ex. [0,1,1] the plane is the b2 x b3 plane in the reciprocal space where the bi are the brillouin primitive vectors
-    initial_grid_spacing: |double| spacing of the k grid before the refinment procedure
-    shift_in_plane : (,2) |array_like| shift of the chosen reciprocal 2D plane with respecet to the crystal coordinates
-    shift_in_space : (,3) |array_like|  shift of the chosen reciprocal 2D plane with respecet to the cartesian coordinates 
-    (this allow to build a 3D k points grid, considering different shifts along the third brillouin primitive vector, if there are no symmetries along this direction to consider)
-    symmetries: |list of lists| a symmetry is a list of 3 elements (the versor is the axis of rotation, while the modulus is the angle)
-    refinment_spacing: |double| initial spacing given to generate the refinment, half of the preceding refinment spacing is considered at each refinment iteration
-    refinment_iteration: |int| number of refinment iterations considered
-    threshold: |double| therhold to recognize a symmetry
-    symmetry_analysis: |boolean| local_refinment_with_symmetry_analysis or not
-    Return
-    -------------
-    "local_refinment_with_symmetry_analysis" not considered:
-        k_points_list: (1,k0*k1) |array_like| k points (kx,ky,kz,w) where w is the weight of the k point
-        parallelograms : (n,4) |array_like| each row is a parallelogram, where the 4 vertices are integers, which correspond to the list positions of the given k points 
-    "local_refinment_with_symmetry_analysis" considered:
-        k_points_list: (1,k0*k1) |array_like| k points (kx,ky,kz,w) where w is the weight of the k point
-        triangles: (n,3) |array_like| each row is a triangle, where the 3 vertices are integers, which correspond to the list positions of the given k points 
-    """
-    normalized_brillouin_primitive_vectors_2d = np.zeros((2, 3),dtype=float)
-    ##print("ecco",precedent_count)
-    if  chosen_plane==[0,0,0]:
-        chosen_plane=[]
-        for i in range(2):
-            normalized_brillouin_primitive_vectors_2d[i] = brillouin_primitive_vectors_2d[i]/(brillouin_primitive_vectors_2d[i]@brillouin_primitive_vectors_2d[i])
-            # redefining the brillouin_primitive_vectors_3d in order to take into account the input brillouin_primitive_vectors_2d
-            for j in range(3):
-                if (brillouin_primitive_vectors_3d[j,:]@brillouin_primitive_vectors_2d[i,:])!=0:
-                    chosen_plane.append(j)
-        count = 0
-        for i in chosen_plane:
-            brillouin_primitive_vectors_3d[i,:]=brillouin_primitive_vectors_2d[count,:]
-            count +=1                    
-    else:
-        #saving the chosen 2d plane 
-        count = 0
-        for i in range(3):
-            if chosen_plane[i] != 0:
-                brillouin_primitive_vectors_2d[count] = brillouin_primitive_vectors_3d[i]
-                normalized_brillouin_primitive_vectors_2d[count] = brillouin_primitive_vectors_2d[count]/(brillouin_primitive_vectors_2d[count]@brillouin_primitive_vectors_2d[count])
-                count += 1
+    ###individuating the 2D plane
+    brillouin_primitive_vectors_2d=np.zeros((2, 3),dtype=float)
+    count = 0
+    for i in range(3):
+        if chosen_plane[i]!=0:
+            brillouin_primitive_vectors_2d[count] = new_brillouin_primitive_vectors_3d[i]
+            count += 1
 
-    #initial 2D grid dimensions
-    print(brillouin_primitive_vectors_2d[1])
-    k0 = int(
-        (brillouin_primitive_vectors_2d[0] @ brillouin_primitive_vectors_2d[0])/ initial_grid_spacing
-        )
-    k1 = int(
-        (brillouin_primitive_vectors_2d[1] @ brillouin_primitive_vectors_2d[1]) / initial_grid_spacing
-        )
-    if k0==0 or k1==0:
+    ###initial 2D grid number of points (considering the grid_spacing input)
+    if grid_spacing!=0.0:
+        k0 = int(
+            (brillouin_primitive_vectors_2d[0] @ brillouin_primitive_vectors_2d[0])/grid_spacing
+            )
+        k1 = int(
+            (brillouin_primitive_vectors_2d[1] @ brillouin_primitive_vectors_2d[1])/grid_spacing
+            )
+        if k0==0 or k1==0:
+            k0=default_gridding
+            k1=default_gridding
+    else:
         k0=default_gridding
         k1=default_gridding
-
-    print("k0 k1",k0,k1,default_gridding)
     initial_weights = 1 / (k0 * k1)
-    #building the initial 2D k points grid
-    k_points_grid = np.zeros((k0*k1,4),dtype=float)
+
+    print("inside 2d generation", brillouin_primitive_vectors_2d,k0,k1)
+
+    ###building the initial 2D k points grid
+    k_points_grid=np.zeros((k0*k1,4),dtype=float)
     for i in range(k0):
         for j in range(k1):
-            k_points_grid[i*k1+j,:3] = (float(i + shift_in_plane[0]) / k0) * (shift_in_space + brillouin_primitive_vectors_2d[0,:]) + (float(j + shift_in_plane[1]) / k1) * (shift_in_space + brillouin_primitive_vectors_2d[1,:])
+            k_points_grid[i*k1+j,:3] =(float(i + shift_in_plane[0]) / k0) * (brillouin_primitive_vectors_2d[0])\
+                       + (float(j + shift_in_plane[1]) / k1) * (brillouin_primitive_vectors_2d[1])
+            k_points_grid[i*k1+j,:3]+=shift_in_space
             k_points_grid[i*k1+j,3] = initial_weights
-            ###print(k_points_grid[i*k1+j,:3])
-    if symmetry_analysis_flag == True:
-        if refinment_spacing==0:
-            refinment_spacing=initial_grid_spacing/2
-
-        #applying the refinment procedure to the 2d k points grid
-        #obtaining a refined list
+    
+    ###symmetry analysis
+    if symmetry_analysis == True:
+        ###if not refinment spacing is given, the one from the building of the k points grid is used
+        if refinment_spacing==0 or refinment_spacing==grid_spacing:
+            refinment_spacing=grid_spacing/2.0
+       
+        ###normalized 2D plange generators, needed in the symmetry-refinment procedure
+        normalized_brillouin_primitive_vectors_2d = np.zeros((2, 3),dtype=float)
+        for i in range(2):
+            normalized_brillouin_primitive_vectors_2d[i]=brillouin_primitive_vectors_2d[i]/(brillouin_primitive_vectors_2d[i]@brillouin_primitive_vectors_2d[i])
+        
+        ###applying the symmetry-refinment procedure to the 2D k points grid
+        ###obtaining a refined list as an output
+        ###loosing the ordering of the k points in the 2D plane
         refined_k_points_list = []
-
         local_refinment_with_symmetry_analysis(
-            refined_k_points_list,
             k_points_grid,
+            refined_k_points_list,
             refinment_spacing,
-            refinment_iteration,
+            refinment_iterations,
             symmetries,
-            threshold,
-            brillouin_primitive_vectors_3d,
+            threshold_symmetry,
+            new_brillouin_primitive_vectors_3d,
             brillouin_primitive_vectors_2d,
             normalized_brillouin_primitive_vectors_2d,
             True
         )
-        #transforming a list into an array-like element
-        k0=int(len(refined_k_points_list)/4)
+        ###transforming the list into an array-like object (calling it list as well)
         refined_k_points_list = np.reshape(refined_k_points_list,(int(len(refined_k_points_list)/4), 4))
-        #applying a triangulation to order the k points
-        #the periodicity of the BZ is properly considered
-        refined_k_points_list,triangles=k_points_triangulation_2d(refined_k_points_list,brillouin_primitive_vectors_2d,0)
-        return refined_k_points_list,triangles
-    else:
-        not_refined_k_points_list = np.zeros((k0*k1,4))
-        parallelograms=[]
-        if np.array_equal(other_brillouin_primitive_vectors_3d,brillouin_primitive_vectors_3d):
-            for i in range(0,k0):
-                for j in range(0,k1):
-                   # print(i,j,k0,k1)
-                    not_refined_k_points_list[i*k1+j][:3]=(float(i + shift_in_plane[0]) / k0) * (shift_in_space + brillouin_primitive_vectors_2d[0]) \
-                       + (float(j + shift_in_plane[1]) / k1) * (shift_in_space + brillouin_primitive_vectors_2d[1])
-                    not_refined_k_points_list[i*k1+j][3]=initial_weights
-                    #periodicity of the BZ
-                    parallelogram=[]
-                    if i<k0-1 and j<k1-1:
-                        parallelogram=[[i*k1+j,0],[(i+1)*k1+j,0],[(i+1)*k1+j+1,0],[i*k1+j+1,0]]
-                    elif i<k0-1 and j==k1-1:
-                        parallelogram=[[i*k1+j,0],[(i+1)*k1+j,0],[(i+1)*k1+0,2],[i*k1+0,2]]
-                    elif i==k0-1 and j<k1-1:
-                        parallelogram=[[i*k1+j,0],[0*k1+j,1],[0*k1+j+1,1],[i*k1+j+1,0]]
-                    else:
-                        parallelogram=[[i*k1+j,0],[0*k1+j,1],[0*k1+0,3],[i*k1+0,2]]
-                    ##print(parallelogram)
-                    #for r in range(4):
-                     #   parallelogram[r][0]+=precedent_count  
-                    ####print(parallelogram)    
-                    parallelograms.append(parallelogram)
-                    ##print(parallelogram[1][0])
-            parallelograms=np.reshape(parallelograms,(int(k0*k1),4,2))
+        #print("prima",refined_k_points_list)
+        refined_k_points_list = np.unique(refined_k_points_list,axis=0)
+        #print("dopo",refined_k_points_list)
+        ###if a covering of the BZ is not requested
+        if covering_BZ == True:
+            ###applying a triangulation procedure to order the k points list 
+            ###the periodicity of the BZ is properly considered in the triangulation procedure
+            refined_k_points_list,triangles=k_points_triangulation_2D(refined_k_points_list,brillouin_primitive_vectors_2d,precedent_count)
+            return refined_k_points_list,triangles
         else:
-            for i in range(0,k0):
-                for j in range(0,k1):
-                   # print(i,j,k0,k1)
-                    not_refined_k_points_list[i*k1+j][:3]=(float(i + shift_in_plane[0]) / k0) * (shift_in_space + brillouin_primitive_vectors_2d[0]) \
-                       + (float(j + shift_in_plane[1]) / k1) * (shift_in_space + brillouin_primitive_vectors_2d[1])
-                    not_refined_k_points_list[i*k1+j][3]=initial_weights
-                    parallelogram=[]
-                    parallelogram=[[i*k1+j,0],[(i+1)*k1+j,0],[(i+1)*k1+j+1,0],[i*k1+j+1,0]]
-                    parallelograms.append(parallelogram)
-                    ##print(parallelogram[1][0])
-            parallelograms=np.reshape(parallelograms,(int(k0*k1),4,2))
-            print("dentro")
-            print(not_refined_k_points_list)
-            ### checking if the points are going over one of the borders of the other_brillouin_primitive_vectors_3d (in the input)
-            not_refined_k_points_list[:,:3],indices=downfold_inside_brillouin_zone(
-                not_refined_k_points_list[:,:3],
-                other_brillouin_primitive_vectors_3d)
-            print(np.shape(not_refined_k_points_list))
-            #print(np.shape(not_refined_k_points_list))
-            print(np.shape(parallelograms))
-            print(np.shape(indices))
-            #print(not_refined_k_points_list[0][3])
-            print(parallelograms)
-
-            for parallelogram in parallelograms:
-                for vertex in parallelogram:
-                    print(vertex)
-                    if indices[vertex[0]][0]==3:
-                        vertex[1]+=-(1+int(vertex[0][1]/k1))*k1
-                        vertex[2]+=-(int(vertex[0][1]%k1)+1)-(1+int(vertex[0][1]/k1))*k1
-                        vertex[3]+=-(int(vertex[0][1]%k1)+1)
-                    elif indices[vertex[0]][0]==2:
-                        vertex[2]+=-(int(vertex[0][1]%k1)+1)
-                        vertex[3]+=-(int(vertex[0][1]%k1)+1)
-                    else:
-                        vertex[1]+=-(1+int(vertex[0][1]/k1))*k1
-                        vertex[2]+=-(1+int(vertex[0][1]/k1))*k1     
-        
-        print(not_refined_k_points_list)      
-        return not_refined_k_points_list,parallelograms
+            refined_k_points_list
+    ###if no symmetry analysis is applied, building parallelograms to cover the BZ
+    else:
+        if covering_BZ == False:
+            k_points_grid
+        else:
+            parallelograms=[]
+            ###in case the downfolding procedure is requested with respect to a different BZ
+            if np.array_equal(new_brillouin_primitive_vectors_3d,old_brillouin_primitive_vectors_3d):
+                for i in range(0,k0):
+                    for j in range(0,k1):
+                        #periodicity of the BZ
+                        parallelogram=[]
+                        if i<k0-1 and j<k1-1:
+                            parallelogram=[[i*k1+j,0],[(i+1)*k1+j,0],[(i+1)*k1+j+1,0],[i*k1+j+1,0]]
+                        elif i<k0-1 and j==k1-1:
+                            parallelogram=[[i*k1+j,0],[(i+1)*k1+j,0],[(i+1)*k1+0,2],[i*k1+0,2]]
+                        elif i==k0-1 and j<k1-1:
+                            parallelogram=[[i*k1+j,0],[0*k1+j,1],[0*k1+j+1,1],[i*k1+j+1,0]]
+                        else:
+                            parallelogram=[[i*k1+j,0],[0*k1+j,1],[0*k1+0,3],[i*k1+0,2]] 
+                        parallelograms.append(parallelogram)
+                parallelograms=np.reshape(parallelograms,(int(k0*k1),4,2))
+            else:
+                k_points_grid[:,:3],indices=downfold_inside_brillouin_zone(
+                    k_points_grid[:,:3],
+                    old_brillouin_primitive_vectors_3d)
+                parallelogram=[]
+                for i in range(k0):
+                    for j in range(k1):
+                        if i<k0-1 and j<k1-1:
+                            parallelogram=[[i*k1+j,indices[i*k1+j]],[(i+1)*k1+j,indices[(i+1)*k1+j]],[(i+1)*k1+j+1,indices[(i+1)*k1+j+1]],[i*k1+j+1,indices[i*k1+j+1]]]
+                            parallelograms.append(parallelogram)
+                parallelograms=np.reshape(parallelograms,(len(parallelograms),4,2))        
+            return k_points_grid,parallelograms
 
 def check_inside_closed_shape_2d(
-    eigenvectors_around_array,
-    refined_k_points_list,
+    border_k_points_list,
+    k_points_list,
     brillouin_primitive_vectors_2d
 ):
-    number_elements_border=eigenvectors_around_array.shape[0]
-    number_elements=refined_k_points_list.shape[0]
-
+    number_elements_border=border_k_points_list.shape[0]
+    number_elements=k_points_list.shape[0]
     k_points_list_projections_border=np.zeros((number_elements_border,2),dtype=float)
-    k_points_list_projections=np.zeros((number_elements_border,2),dtype=float)
+    k_points_list_projections=np.zeros((number_elements,2),dtype=float)
 
     origin=np.zeros(3,dtype=float)
-
-    #calculating the projections of the different k points on the primitive vectors
+    ### calculating the projections of the different k points on the primitive vectors
     for i in range(number_elements_border):
         for j in range(2):
-            k_points_list_projections_border[i,j]=(eigenvectors_around_array[i,:3]-origin)@brillouin_primitive_vectors_2d[j,:]
+            k_points_list_projections_border[i,j]=(border_k_points_list[i,:3]-origin)@brillouin_primitive_vectors_2d[j,:]
     for i in range(number_elements):
         for j in range(2):
-            k_points_list_projections[i,j]=(refined_k_points_list[i,:3]-origin)@brillouin_primitive_vectors_2d[j,:]
+            k_points_list_projections[i,j]=(k_points_list[i,:3]-origin)@brillouin_primitive_vectors_2d[j,:]
     
-    #Point Inclusion in Polygon Test W. Randolph Franklin (WRF) 
+    #point inclusion in polygon test W. Randolph Franklin (WRF) 
     def pnpoly(nvert,vert,test):
         i=0
         c=1
@@ -956,78 +840,65 @@ def check_inside_closed_shape_2d(
             for j in range(i+1,nvert-1):
                 if ( ((vert[i,1]>test[1]) != (vert[j,1]>test[1])) and (test[0]<(vert[j,0]-vert[i,0])*(test[1]-vert[i,1])/(vert[j,1]-vert[i,1])+vert[i,0]) ):    
                     c = -c
+            i+=1
         return c
+    
     indices=[]
     for i in range(number_elements):
         if pnpoly(number_elements_border,k_points_list_projections_border,k_points_list_projections[i])<0:
             indices.append(i)
-    del refined_k_points_list[indices]
-    return refined_k_points_list
+    
+    ##if len(indices)!=0:
+    ##    count=0
+    ##    for i in range(len(indices)):
+    ##        k_points_list=np.delete(k_points_list,i-count,axis=0)
+    ##        count+=1
 
-def dynamical_refinment_little_paths_2d(
+    return k_points_list
+
+def dynamical_refinment_little_paths_2D(
         little_paths,
+        k_points_list,
         position_little_paths_to_refine,
         number_vertices,
-        all_k_points_list,
-        refinment_iteration,
+        refinment_iterations,
         symmetries,
-        threshold,
+        threshold_symmetry,
         brillouin_primitive_vectors_3d,
-        brillouin_primitive_vectors_2d,
         chosen_plane,
+        brillouin_primitive_vectors_2d,
         normalized_brillouin_primitive_vectors_2d,
-        epsilon
+        threshold_minimal_refinment=0.00000000001,
+        default_gridding=100
 ):
-    r"""
-    some k points of the "all k points list" are selected (the ones in the selected little paths "position_littles_paths_to_refine"), 
-    and a local refinment is applied to them, in case of triangles (number_vertices=3) a refinment with symmetry analysis is applied,
-    in case of parallelograms (number_vertices=4) a refinment without symmetry analysis is applied 
-    points on the border or nearby the border of the BZ are properly considered
-
-    a refinmente without symmetry analysis is done in the case of parallelograms 
-
-    Parameters
-    ---------
-    little_paths: (s,number_vertices) |array_like| (v1,v2,v3...) int (v=vertex)
-    all_k_points_list: (n,4) |array_like| (kx,ky,kz,w)
-    positions_little_paths_to_refine: |list| (int) positions of the little paths to which the refinment procedure is applied
-    number_vertices: |int| distinguishing between traingles and parallelograms
-
-    (parameters for the refinment...)
-
-    Returns
-    ---------
-    new_k_points_list: (n,4) |array_like| (kx,ky,kz,w)
-    added_little_paths: (,3) |array_like| (v1,v2,v3)
-    """
-    ###print(all_k_points_list)
-    new_all_k_points_list=all_k_points_list
+    new_k_points_list=k_points_list
+    counting_offset=len(k_points_list)
     only_new_little_paths=[]
     number_little_paths=len(little_paths)
     eigenspaces_little_paths_to_refine=[[i] for i in position_little_paths_to_refine]
+    print("beginning: ",eigenspaces_little_paths_to_refine)
     flag_checking_indipendence=True
-    print(position_little_paths_to_refine)
     while flag_checking_indipendence == True:
         number_eigenspaces=len(eigenspaces_little_paths_to_refine)
         number_total_eigenvectors=0
-        number_total_around=0
+        number_total_little_paths_around=0
         little_paths_around_each_eigenspace=[]
-        print("eigenspaces",eigenspaces_little_paths_to_refine)
+        print(eigenspaces_little_paths_to_refine)
         for eigenspace in eigenspaces_little_paths_to_refine:
             print(eigenspace)
             number_eigenvectors=len(eigenspace)
             number_total_eigenvectors+=number_eigenvectors
-            # considering the to-refine little paths in the selected eigenspace
+            ### considering the to-refine little paths in the selected eigenspace
             eigenvectors = little_paths[eigenspace]
-            # the little paths around each eigenvector are associated to it
+            ### the little paths around each eigenvector are associated to it
             little_paths_around_each_eigenvector=[]
             for i in range(number_eigenvectors):
                 little_paths_around_each_eigenvector.extend([r for r in range(number_little_paths) for j in little_paths[r][:,0] if j in eigenvectors[i][:,0] and r not in eigenspace])
             little_paths_around_each_eigenvector=list(np.unique(little_paths_around_each_eigenvector))
-            number_total_around+=len(little_paths_around_each_eigenvector)
+            number_total_little_paths_around+=len(little_paths_around_each_eigenvector)
             little_paths_around_each_eigenspace.append(little_paths_around_each_eigenvector)
-        # checking if two eigenspaces are in reality the same eigenspace:
-        # between the "little_paths_around" of one of the two eigenspaces the index of one of the eigenvectors of the other eigenspace appear
+        ### checking if two eigenspaces are in reality the same eigenspace
+        ### between the "little_paths_around" of one of the two eigenspaces the index of one of the eigenvectors of the other eigenspace appears
         check_degeneracy = np.zeros((number_eigenspaces, number_eigenspaces), dtype=bool)
         any_degeneracy = 0
         for i in range(number_eigenspaces-1):
@@ -1040,272 +911,380 @@ def dynamical_refinment_little_paths_2d(
                     else:
                         check_degeneracy[i,j]=False
         if any_degeneracy!=0:
-            #unifying the eigenspaces properly
-            #after having written them in an array form to facilitate it (the unification)
-            #unifying as well the "little paths around"
+            ## unifying the eigenspaces properly
+            ## after having written them in an array form to facilitate the unification
+            ## unifying as well the "little paths around"
             eigenspaces_little_paths_to_refine_array=np.zeros((number_total_eigenvectors,2),dtype=int)
-            little_paths_around_each_eigenspace_array=np.zeros((number_total_around,2),dtype=int)
+            little_paths_around_each_eigenspace_array=np.zeros((number_total_little_paths_around,2),dtype=int)
             count1=0
             count2=0
-            count4=0
+            count3=0
             for eigenspace in eigenspaces_little_paths_to_refine:
-                print("degeneracy checking",eigenspace)
-                count3=0
                 for eigenvector in eigenspace:
                     eigenspaces_little_paths_to_refine_array[count1,0]=eigenvector
                     eigenspaces_little_paths_to_refine_array[count1,1]=count2
                     count1+=1
                 for around in little_paths_around_each_eigenspace[count2]:
-                    little_paths_around_each_eigenspace_array[count4,0]=around
-                    little_paths_around_each_eigenspace_array[count4,1]=count2
-                    count4+=1
-                count3+=1
+                    little_paths_around_each_eigenspace_array[count3,0]=around
+                    little_paths_around_each_eigenspace_array[count3,1]=count2
+                    count3+=1
                 count2+=1
-            count=0
-            print("before degeneracy",eigenspaces_little_paths_to_refine_array)
-            print(check_degeneracy)
             for i in range(number_eigenspaces-1):
                 for j in range(i+1,number_eigenspaces):
                     if check_degeneracy[i][j]==True:
                         min_value=min(i,j)
                         max_value=max(i,j)
                         eigenvectors_positions=[r for r in range(number_total_eigenvectors) if eigenspaces_little_paths_to_refine_array[r,1]==max_value]
-                        around_positions=[r for r in range(number_total_around) if little_paths_around_each_eigenspace_array[r,1]==max_value]
+                        around_positions=[r for r in range(number_total_little_paths_around) if little_paths_around_each_eigenspace_array[r,1]==max_value]
                         eigenspaces_little_paths_to_refine_array[eigenvectors_positions,1]=min_value
                         little_paths_around_each_eigenspace_array[around_positions,1]=min_value
-            print("after degeneracy",eigenspaces_little_paths_to_refine_array)
-            # redefining eigenspaces in order to take into account the found degeneracies
+            ### redefining eigenspaces in order to take into account the found degeneracies
             values_eigenspaces=list(np.unique(eigenspaces_little_paths_to_refine_array[:,1]))
             eigenspaces_little_paths_to_refine=[]
             little_paths_around_each_eigenspace=[]
             for value in values_eigenspaces:
                 eigenspaces_little_paths_to_refine.append([eigenspaces_little_paths_to_refine_array[r,0] for r in range(number_total_eigenvectors) if eigenspaces_little_paths_to_refine_array[r,1]==value])
-                little_paths_around_each_eigenspace.append([little_paths_around_each_eigenspace_array[r,0] for r in range(number_total_around) if little_paths_around_each_eigenspace_array[r,1]==value])
-            print("proper renaming",eigenspaces_little_paths_to_refine,little_paths_around_each_eigenspace)
-            # eliminating in each eigenspace the presence of eigenvectors in other eigenvectors little paths around
+                little_paths_around_each_eigenspace.append([little_paths_around_each_eigenspace_array[r,0] for r in range(number_total_little_paths_around) if little_paths_around_each_eigenspace_array[r,1]==value])
+            ### eliminating in each eigenspace the presence of eigenvectors in other eigenvectors little paths around
             number_eigenspaces=len(eigenspaces_little_paths_to_refine)
-            print(number_eigenspaces)
             for i in range(number_eigenspaces):
-                print(i,little_paths_around_each_eigenspace[i])
                 indices=[]
-                count6=0
+                count4=0
                 for r in little_paths_around_each_eigenspace[i]:
                     for j in eigenspaces_little_paths_to_refine[i]:
                         if j == r:
-                            indices.append(count6)
-                    count6+=1
+                            indices.append(count4)
+                    count4+=1
                 indices.sort()
-                print("indices",indices)
                 if len(indices)!=0:
-                    count5=0
+                    count4=0
                     for index in indices:
-                        print(index)
-                        del little_paths_around_each_eigenspace[i][index-count5]
-                        count5+=1
-                print("after erasing",little_paths_around_each_eigenspace)
-            # checking if two eigenspaces have one little path around in common 
-            # this in common little path is inserted into the to refine little paths and the procedure is repeated
-            new_little_paths_to_refine=[]
-            for i in range(number_eigenspaces-1):
-                for j in range(i+1,number_eigenspaces):
-                    for element in little_paths_around_each_eigenspace[i]:
-                        if element in little_paths_around_each_eigenspace[j]:
-                            new_little_paths_to_refine.append(element)
-            if len(new_little_paths_to_refine)==0:
-                flag_checking_indipendence=True
-            else:
-                new_little_paths_to_refine=list(np.unique(new_little_paths_to_refine))
-                for s in new_little_paths_to_refine:
-                    eigenspaces_little_paths_to_refine.append([s])
+                        del little_paths_around_each_eigenspace[i][index-count4]
+                        count4+=1
+        no_regularization=0
+        ### checking if two eigenspaces have one little path around in common 
+        ### this in common little path is inserted into the to refine little paths and the procedure is repeated
+        list_tmp=[]
+        for i in range(number_eigenspaces-1):
+            for j in range(i+1,number_eigenspaces):
+                for element in little_paths_around_each_eigenspace[i]:
+                    if element in little_paths_around_each_eigenspace[j]:
+                        list_tmp.append(element)
+        if len(list_tmp)!=0:
+            no_regularization+=1
+            list_tmp=list(np.unique(list_tmp))
+            for s in list_tmp:
+                eigenspaces_little_paths_to_refine.append([s])
+        ### symmetrizing the refinment region (mainly for the number_vertices = 4 case)
+        ### check for each eigenspace, if two eigenvectors have one common near-by little path
+        ### that near-by little path is considered as a selected little-path if it has at least two vertices in common with each eigenvector
+        for eigenspace in eigenspaces_little_paths_to_refine:
+            number_eigenvectors=len(eigenspace)
+            if number_eigenvectors>1:
+                near_by_little_paths_2_vertices=np.empty(number_eigenvectors,dtype=object)
+                list_tmp=[]
+                count7=0
+                for eigenvector in eigenspace:
+                    for vertex_2 in little_paths[eigenvector]:
+                        for r in range(number_little_paths):
+                            count8=0
+                            for vertex_1 in little_paths[r]:
+                                if vertex_1[0]==vertex_2[0] and vertex_1[1]==vertex_2[1]:
+                                    count8+=1
+                            if count8>=2:
+                                list_tmp.append(r)
+                near_by_little_paths_2_vertices[count7]=np.unique(list_tmp)
+                if len(near_by_little_paths_2_vertices[count7])!=0:
+                    for i in range(0,number_eigenvectors-1):
+                        for j in range(i+1,number_eigenvectors):
+                            for r in near_by_little_paths_2_vertices[i]:
+                                for s in near_by_little_paths_2_vertices[j]:
+                                    if s==r:
+                                        no_regularization+=1
+                                        eigenspaces_little_paths_to_refine.append([s])
+                count7+=1
+        if no_regularization!=0 or any_degeneracy!=0:
+            flag_checking_indipendence=True
+            print("degeneracies and regularizations detected: ",any_degeneracy,no_regularization)
         else:
             flag_checking_indipendence=False
-        print("finish cycle",eigenspaces_little_paths_to_refine,little_paths_around_each_eigenspace)
-    print("Little paths to refine:", eigenspaces_little_paths_to_refine)
-    print("Little paths around the ones to refine:", little_paths_around_each_eigenspace)
-    #substituting to the little_paths_positions, the k points positions
-    print("in the middle",new_all_k_points_list)
+        print("cycling",eigenspaces_little_paths_to_refine,little_paths_around_each_eigenspace)
+    print("little paths to refine:",eigenspaces_little_paths_to_refine)
+    print("little paths around the ones to refine:",little_paths_around_each_eigenspace)
+    ### substituting to the little paths positions, the k points positions in the list of k points
     number_eigenspaces=len(eigenspaces_little_paths_to_refine)
     eigenspaces_k_points_to_refine=[]
     eigenspaces_k_points_around=[]
     for eigenspace in eigenspaces_little_paths_to_refine:
         eigenspaces_k_points_to_refine.append(np.unique([vertex for eigenvector in eigenspace for vertex in little_paths[eigenvector]],axis=0))
-        for eigenvector in eigenspace:
-            print("k point paths to refine",little_paths[eigenvector])
-    for eigenspace in little_paths_around_each_eigenspace:
-        eigenspaces_k_points_around.append(np.unique([vertex for eigenvector in eigenspace for vertex in little_paths[eigenvector]],axis=0))    
-    print("K points to refine:", np.shape(eigenspaces_k_points_to_refine))
-    print(eigenspaces_k_points_to_refine)
-    print("K points around the ones to refine:", np.shape(eigenspaces_k_points_around))
-    print(eigenspaces_k_points_around)
-    # procede to a refinement of the k points constituing the to-refine little paths
+    ### erasing from the k points around the one inside the refinment region
+    count5=0
+    for eigenspace_2 in little_paths_around_each_eigenspace:
+        list_tmp=[]
+        for eigenvector_2 in eigenspace_2:
+            for vertex_2 in little_paths[eigenvector_2]:
+                count6=0
+                for eigenvector_1 in eigenspaces_little_paths_to_refine[count5]:
+                    for vertex_1 in little_paths[eigenvector_1]:
+                        if vertex_1[0]==vertex_2[0]:
+                            count6+=1
+                if count6==0:
+                    list_tmp.extend([list(vertex_2)])
+        eigenspaces_k_points_around.append(np.unique(np.asarray(list_tmp),axis=0))
+        count5+=1
+    print("k points to refine:",eigenspaces_k_points_to_refine)
+    print("k points around the ones to refine:",eigenspaces_k_points_around)
+    ### procede to a refinement of the k points constituing the to-refine little paths
     if number_vertices==3:
-        all_new_triangles=[]
-        all_k_points_list_tmp=[]
+        k_points_list_tmp=[]
+        only_new_little_paths=[]
         # finding the minimum distance between the to-refine k points and the border k points
         for i in range(number_eigenspaces):
             number_eigenvectors=len(eigenspaces_k_points_to_refine[i])
-            eigenvectors_array=np.zeros((number_eigenvectors,3),dtype=float)
+            eigenvectors_array=np.zeros((number_eigenvectors,4),dtype=float)
             number_eigenvectors_around=len(eigenspaces_k_points_around[i])
-            eigenvectors_around_array=np.zeros((number_eigenvectors_around,3),dtype=float)
+            eigenvectors_around_array=np.zeros((number_eigenvectors_around,4),dtype=float)
             for j in range(number_eigenvectors):
                 if eigenspaces_k_points_to_refine[i][j,1]==0:
-                    eigenvectors_array[j]=all_k_points_list[eigenspaces_k_points_to_refine[i][j,0]]
+                    eigenvectors_array[j,:3]=k_points_list[eigenspaces_k_points_to_refine[i][j,0],:3]
                 elif eigenspaces_k_points_to_refine[i][j,1]==1:
-                    eigenvectors_array[j]=all_k_points_list[eigenspaces_k_points_to_refine[i][j,0]]+brillouin_primitive_vectors_2d[0]
+                    eigenvectors_array[j,:3]=k_points_list[eigenspaces_k_points_to_refine[i][j,0],:3]+brillouin_primitive_vectors_2d[0]
+                elif eigenspaces_k_points_to_refine[i][j,1]==2:
+                    eigenvectors_array[j,:3]=k_points_list[eigenspaces_k_points_to_refine[i][j,0],:3]+brillouin_primitive_vectors_2d[1]
                 else:
-                    eigenvectors_array[j]=all_k_points_list[eigenspaces_k_points_to_refine[i][j,0]]+brillouin_primitive_vectors_2d[1]
-            minimal_distance = (cdist(eigenvectors_array-eigenvectors_around_array)).min()
-            refined_k_points_list=[]
+                    eigenvectors_array[j,:3]=k_points_list[eigenspaces_k_points_to_refine[i][j,0],:3]+brillouin_primitive_vectors_2d[1]+brillouin_primitive_vectors_2d[0]
+                eigenvectors_array[j,3]=k_points_list[eigenspaces_k_points_to_refine[i][j,0],3]
+            border_indices=[]
+            for j in range(number_eigenvectors_around):
+                if eigenspaces_k_points_around[i][j,1]==0:
+                    eigenvectors_around_array[j,:3]=k_points_list[eigenspaces_k_points_around[i][j,0],:3]
+                elif eigenspaces_k_points_around[i][j,1]==1:
+                    eigenvectors_around_array[j,:3]=k_points_list[eigenspaces_k_points_around[i][j,0],:3]+brillouin_primitive_vectors_2d[0]
+                elif eigenspaces_k_points_around[i][j,1]==2:
+                    eigenvectors_around_array[j,:3]=k_points_list[eigenspaces_k_points_around[i][j,0],:3]+brillouin_primitive_vectors_2d[1]
+                else:
+                    eigenvectors_around_array[j,:3]=k_points_list[eigenspaces_k_points_around[i][j,0],:3]+brillouin_primitive_vectors_2d[1]+brillouin_primitive_vectors_2d[0]
+                eigenvectors_around_array[j,3]=k_points_list[eigenspaces_k_points_around[i][j,0],3]
+                border_indices.append(eigenspaces_k_points_around[i][j,0])
+            distance=np.zeros(number_eigenvectors_around*number_eigenvectors,dtype=float)
+            for s in range(number_eigenvectors):
+                for t in range(number_eigenvectors_around):
+                    distance[s*number_eigenvectors_around+t]=np.linalg.norm(eigenvectors_array[s,:3]-eigenvectors_around_array[t,:3],2)
+                    print(eigenvectors_array[s,:3],eigenvectors_around_array[t,:3],s,t)
+            minimal_distance = np.min(distance)
+            try:
+                1./minimal_distance
+            except ZeroDivisionError:
+                print("Acthung distance: ",minimal_distance)
+            refined_k_points_list_tmp=[]
             local_refinment_with_symmetry_analysis(
-                refined_k_points_list,
-                eigenvectors_array,
-                minimal_distance,
-                refinment_iteration,
-                symmetries,
-                threshold,
-                brillouin_primitive_vectors_3d,
-                brillouin_primitive_vectors_2d,
-                normalized_brillouin_primitive_vectors_2d,
-                False
-            )
-            # transforming from a list to array_like elements
-            refined_k_points_list = np.reshape(refined_k_points_list,(int(len(refined_k_points_list)/4), 4))
-            # check if the points are inside the border 
+                    eigenvectors_array,
+                    refined_k_points_list_tmp,
+                    minimal_distance,
+                    refinment_iterations,
+                    symmetries,
+                    threshold_symmetry,
+                    brillouin_primitive_vectors_3d,
+                    brillouin_primitive_vectors_2d,
+                    normalized_brillouin_primitive_vectors_2d,
+                    False
+                    )
+            ###transforming the list into an array-like object (calling it list as well)
+            refined_k_points_list_tmp = np.reshape(refined_k_points_list_tmp,(int(len(refined_k_points_list_tmp)/4), 4))
+            ###check if the points are inside the border
             refined_k_points_list = check_inside_closed_shape_2d(
                                         eigenvectors_around_array,
-                                        refined_k_points_list,
+                                        refined_k_points_list_tmp,
                                         brillouin_primitive_vectors_2d)
-            # triangulation
-            refined_k_points_list,added_triangles=k_points_triangulation_2d(
-                                                    refined_k_points_list,
+            ### triangulation
+            refined_k_points_list_tmp = np.append(refined_k_points_list,eigenvectors_around_array,axis=0)
+            refined_k_points_list_tmp,triangles=k_points_triangulation_2D(
+                                                    refined_k_points_list_tmp,
                                                     brillouin_primitive_vectors_2d,
-                                                    count)
-            count+=len(refined_k_points_list)
-            all_k_points_list_tmp.append(refined_k_points_list)
-            all_new_triangles.append(added_triangles)
-        all_k_points_list.append(all_k_points_list_tmp)
-        return all_k_points_list, added_triangles
+                                                    counting_offset,
+                                                    border_indices)
+            ### checking border BZ
+            refined_k_points_list_tmp[:,:3],indices=downfold_inside_brillouin_zone(
+                                                    refined_k_points_list_tmp[:,:3],
+                                                    brillouin_primitive_vectors_3d,
+                                                    )
+            for triangle in triangles:
+                for vertex in triangle:
+                    index=vertex[0]
+                    print(index,counting_offset)
+                    if index >= counting_offset:
+                        print("dentro")
+                        vertex[1]=indices[index-counting_offset]
+            ### adding the new k points and triangles to the list of output
+                
+            counting_offset+=len(refined_k_points_list)
+            k_points_list_tmp.extend(refined_k_points_list)
+            only_new_little_paths.extend(triangles)
+
+        ### ordering the lists
+        only_new_little_paths=np.reshape(only_new_little_paths,(int(len(only_new_little_paths)),3,2))
+        new_k_points_list=np.append(new_k_points_list,k_points_list_tmp,axis=0)
     else:
-        print("dentro2")
-        # here it is sufficient to find for each eigenspace the extremal k points
-        # these k points are then used to draw a simil-BZ, which is refined using the grid-generation 2d methods
-        precedent_count=len(all_k_points_list)
-        ###print("ecco2",precedent_count)
-       # print(np.shape(eigenspaces_k_points_around))
+        ### here it is sufficient to find for each eigenspace the extremal k points
+        ### these k points are then used to draw a simil-BZ, which is refined using the grid-generation 2D methods
+        ### first up-folding the k points 
         for i in range(number_eigenspaces):
             number_eigenvectors_around=len(eigenspaces_k_points_around[i])
-            #print(number_eigenvectors_around)
             eigenvectors_around_array=np.zeros((number_eigenvectors_around,4),dtype=float)
-            #print(eigenspaces_k_points_around[i])
             for j in range(number_eigenvectors_around):
-                print(all_k_points_list[eigenspaces_k_points_around[i][j][0]])
                 if eigenspaces_k_points_around[i][j][1]==3:
-                    eigenvectors_around_array[j,:3]=all_k_points_list[eigenspaces_k_points_around[i][j][0],:3]+brillouin_primitive_vectors_2d[1]+brillouin_primitive_vectors_2d[0]
+                    eigenvectors_around_array[j,:3]=k_points_list[eigenspaces_k_points_around[i][j][0],:3]+brillouin_primitive_vectors_2d[1]+brillouin_primitive_vectors_2d[0]
                 elif eigenspaces_k_points_around[i][j][1]==2:
-                    eigenvectors_around_array[j,:3]=all_k_points_list[eigenspaces_k_points_around[i][j][0],:3]+brillouin_primitive_vectors_2d[1]
+                    eigenvectors_around_array[j,:3]=k_points_list[eigenspaces_k_points_around[i][j][0],:3]+brillouin_primitive_vectors_2d[1]
                 elif eigenspaces_k_points_around[i][j][1]==1:
-                    eigenvectors_around_array[j,:3]=all_k_points_list[eigenspaces_k_points_around[i][j][0],:3]+brillouin_primitive_vectors_2d[0]
+                    eigenvectors_around_array[j,:3]=k_points_list[eigenspaces_k_points_around[i][j][0],:3]+brillouin_primitive_vectors_2d[0]
                 else:
-                    eigenvectors_around_array[j]=all_k_points_list[eigenspaces_k_points_around[i][j][0]]
-                eigenvectors_around_array[j,3]=all_k_points_list[eigenspaces_k_points_around[i][j][0],3]
-            #print(eigenvectors_around_array)
-            #calculating the projections of the different k points on the primitive vectors
-            print(brillouin_primitive_vectors_2d,eigenvectors_around_array)
+                    eigenvectors_around_array[j]=k_points_list[eigenspaces_k_points_around[i][j][0]]
+                eigenvectors_around_array[j,3]=k_points_list[eigenspaces_k_points_around[i][j][0],3]  
+            ### calculating the projections of the different k points on the primitive vectors
             k_points_list_projections_border=np.zeros((number_eigenvectors_around,2),dtype=float)
-            for i in range(number_eigenvectors_around):
+            k_points_list_projections_border_sum=np.zeros((number_eigenvectors_around),dtype=float)
+            for r in range(number_eigenvectors_around):
                 for j in range(2):
-                    k_points_list_projections_border[i,j]=(eigenvectors_around_array[i,:3])@brillouin_primitive_vectors_2d[j,:]
-            #transforming into polar coordinates x,y -> rho,theta
-            rho=list(map(lambda x,y: math.sqrt(x**2+y**2),k_points_list_projections_border[:,0],k_points_list_projections_border[:,1]))
-            rho=np.reshape(rho,(number_eigenvectors_around,1))
-            ij_origin=np.argmin(rho)
-            print(rho)
-            origin=eigenvectors_around_array[ij_origin,:3]
-            print("origine:", origin)
-            #defining the projections respect to the new origin
-            for i in range(number_eigenvectors_around):
+                    k_points_list_projections_border[r,j]=eigenvectors_around_array[r,:3]@brillouin_primitive_vectors_2d[j,:]
+                k_points_list_projections_border_sum[r]=np.sum(k_points_list_projections_border[r])
+            ### sorting in order to find the simil-BZ
+            origin_ij=np.argmin(k_points_list_projections_border_sum)
+            origin=eigenvectors_around_array[origin_ij,:3]
+            for s in range(number_eigenvectors_around):
                 for j in range(2):
-                    k_points_list_projections_border[i,j]=((eigenvectors_around_array[i,:3])-origin)@brillouin_primitive_vectors_2d[j,:]
-            #transforming into polar coordinates x,y -> rho,theta
-            rho=list(map(lambda x,y: math.sqrt(x**2+y**2),k_points_list_projections_border[:,0],k_points_list_projections_border[:,1]))
-            rho=np.reshape(rho,(number_eigenvectors_around,1))
-            theta=list(map(lambda x,y: math.atan(y/(x+epsilon)),k_points_list_projections_border[:,0],k_points_list_projections_border[:,1]))
-            for i in range(len(theta)):
-                if theta[i]<0:
-                    theta[i]=-theta[i]+np.pi/2
-            theta=np.reshape(theta,(number_eigenvectors_around,1))
-            #print("theta ",theta)
-            #print("rho",rho)
-            theta_0=np.min(theta)
-            ij_a0=[i for i in range(number_eigenvectors_around) if np.isclose(theta[i],theta_0,atol=epsilon)]
-            rho_0=np.max(rho[ij_a0])
-            j_a0=np.min([i for i in range(number_eigenvectors_around) if np.isclose(rho_0,rho[i],atol=epsilon) and i in ij_a0])
-            #i_a0=[i for i in range(number_eigenvectors_around) if np.isclose(np.min(rho[ij_a0]),rho[i])]
-            theta_1=np.max(theta)
-            ij_a1=[i for i in range(number_eigenvectors_around) if np.isclose(theta[i],theta_1,atol=epsilon)]
-            rho_1=np.max(rho[ij_a1])
-            j_a1=np.min([i for i in range(number_eigenvectors_around) if np.isclose(rho_1,rho[i],atol=epsilon) and i in ij_a1])
-            #i_a1=[i for i in range(number_eigenvectors_around) if np.isclose(np.min(rho[ij_a1]),rho[i],atol=epsilon)]
-            #print(theta_0,theta_1)
-            #print(rho_0,rho_1)
-            #print("indices: ",j_a0,j_a1)
-            #print(origin,eigenvectors_around_array[j_a0,:3],eigenvectors_around_array[j_a1,:3])
-            #finding the new bordening
+                    k_points_list_projections_border[s,j]=(eigenvectors_around_array[s,:3]-origin)@brillouin_primitive_vectors_2d[j,:]
+            ##print(k_points_list_projections_border)
+            ### finding the bottom-right point
+            max_0=k_points_list_projections_border[0,0]
+            min_1=k_points_list_projections_border[0,1]
+            position_pair_0=0
+            for t in range(number_eigenvectors_around):
+                if k_points_list_projections_border[t,0]>=max_0 and k_points_list_projections_border[t,1]<=min_1:
+                    max_0=k_points_list_projections_border[t,0]
+                    min_1=k_points_list_projections_border[t,1]
+                    position_pair_0=t
+                    #print(position_pair_0)
+            ### finding the top-left point
+            min_0=k_points_list_projections_border[0,0]
+            max_1=k_points_list_projections_border[0,1]
+            position_pair_1=0
+            for t in range(number_eigenvectors_around):
+                if k_points_list_projections_border[t,0]<=min_0 and  k_points_list_projections_border[t,1]>=max_1:
+                    max_0=k_points_list_projections_border[t,0]
+                    min_1=k_points_list_projections_border[t,1]
+                    position_pair_1=t
+            print("extremal positions",position_pair_0,position_pair_1,origin_ij)
+            ### defining the new bordening
             new_brillouin_primitive_vectors_2d=np.zeros((2,3),dtype=float)
-            new_brillouin_primitive_vectors_2d[0]=eigenvectors_around_array[j_a0,:3]-origin
-            new_brillouin_primitive_vectors_2d[1]=eigenvectors_around_array[j_a1,:3]-origin
-            #print(new_brillouin_primitive_vectors_2d)
+            new_brillouin_primitive_vectors_2d[0]=eigenvectors_around_array[position_pair_0,:3]-origin
+            new_brillouin_primitive_vectors_2d[1]=eigenvectors_around_array[position_pair_1,:3]-origin
             new_brillouin_primitive_vectors_3d=np.zeros((3,3),dtype=float)
-            new_brillouin_primitive_vectors_3d[2,:]=brillouin_primitive_vectors_3d[[i for i in range(3) if chosen_plane[i]==0],:]
-            for i in range(2):
-                new_brillouin_primitive_vectors_3d[i,:]=new_brillouin_primitive_vectors_2d[i,:]
-            
-            #finding the minimal grid spacing
+            count = 0
+            for s in range(3):
+                if chosen_plane[s] != 0:
+                    new_brillouin_primitive_vectors_3d[s] = new_brillouin_primitive_vectors_2d[count]
+                    count += 1
+                else:
+                    new_brillouin_primitive_vectors_3d[s] = brillouin_primitive_vectors_3d[s]
+            print("new brillouin zone \n", new_brillouin_primitive_vectors_2d)
+            ### finding the minimal grid spacing
             max_value=np.linalg.norm(new_brillouin_primitive_vectors_2d[0]+new_brillouin_primitive_vectors_2d[1],2)
             moduli=np.zeros(number_eigenvectors_around,dtype=float)
-            for i in range(number_eigenvectors_around):
-                moduli[i]=np.linalg.norm(eigenvectors_around_array[i,:3]-origin,2)
-                if moduli[i]<epsilon:
-                    moduli[i]=max_value
-            initial_grid_spacing=np.min(moduli)/2
-            #print(initial_grid_spacing)
-            print("input gridding",new_brillouin_primitive_vectors_3d,
-              initial_grid_spacing,
-              chosen_plane,
-              new_brillouin_primitive_vectors_2d,
-              origin)
-            
-            new_not_refined_k_points_list,new_parallelograms=k_points_generator_2D(
-                new_brillouin_primitive_vectors_3d,
-                brillouin_primitive_vectors_3d,
-                initial_grid_spacing,
-                chosen_plane,
-                new_brillouin_primitive_vectors_2d,
-                100,
-                False,
-                0,
-                [[0,0,0]],
-                0.0,
-                0,
-                [0,0],
-                origin,
-                precedent_count
-                )
+            for s in range(number_eigenvectors_around):
+                moduli[s]=np.linalg.norm(eigenvectors_around_array[s,:3]-origin,2)
+                if moduli[s] < threshold_minimal_refinment:
+                    moduli[s]=max_value
+            initial_grid_spacing=np.min(moduli)/6
+            print("input 2d generation",initial_grid_spacing,counting_offset,origin)
+            new_k_points_list_tmp,parallelograms=k_points_generator_2D(
+                                                    new_brillouin_primitive_vectors_3d,
+                                                    brillouin_primitive_vectors_3d,
+                                                    chosen_plane,
+                                                    initial_grid_spacing,
+                                                    True,
+                                                    default_gridding,
+                                                    False,
+                                                    0,
+                                                    0,
+                                                    [[0,0,0]],
+                                                    0.0,
+                                                    [0,0],
+                                                    origin,
+                                                    counting_offset
+                                                    )
+            ### saving the parallelograms           
             for parallelogram in parallelograms:
                 for vertex in parallelogram:
-                    vertex[1]+=precedent_count
-           ### print(new_all_k_points_list)
-           #### print(new_not_refined_k_points_list)
-            precedent_count+=len(new_not_refined_k_points_list)
-            new_all_k_points_list=np.append(new_all_k_points_list,new_not_refined_k_points_list,axis=0)
-            only_new_little_paths.extend(new_parallelograms)
+                    vertex[0]+=counting_offset
+            counting_offset+=len(new_k_points_list_tmp)
+            new_k_points_list=np.append(new_k_points_list,new_k_points_list_tmp,axis=0)
+            only_new_little_paths.extend(parallelograms)
+            print("new little paths", np.shape(only_new_little_paths))
+            print("k points list", np.shape(new_k_points_list))
+    print("prima",only_new_little_paths)
+    return new_k_points_list,only_new_little_paths
 
-        print(np.shape(only_new_little_paths))
-        ##only_new_little_paths=np.reshape(only_new_little_paths,(int(len(only_new_little_paths)/3),3))
-        
-        return new_all_k_points_list,only_new_little_paths
+def printing_covering_BZ_2D(
+    brillouin_primitive_vectors_3d,
+    chosen_plane,
+    k_points_list,
+    little_paths,
+    number_vertices
+    ):
+    
+    brillouin_primitive_vectors_2d=np.zeros((2,3),dtype=float)
+    count = 0
+    for i in range(3):
+        if chosen_plane[i]!=0:
+            brillouin_primitive_vectors_2d[count] = brillouin_primitive_vectors_3d[i]
+            count += 1
+    
+    number_little_paths=len(little_paths)
+    little_path=[0]*number_vertices
+    ### substituing to the vertices the k points coordinates
+    new_little_paths=[]
+    for i in range(number_little_paths):
+        indx_1=[little_paths[i][r][0] for r in range(number_vertices)]
+        indx_2=[little_paths[i][r][1] for r in range(number_vertices)]
+        count=0
+        for indx in indx_1:
+            little_path[count]=[k_points_list[indx,s] for s in range(3)]
+            if indx_2[count]!=0:
+                for s in range(3):
+                    if indx_2[count]==1:
+                        little_path[count][s]+=brillouin_primitive_vectors_2d[0,s]
+                    elif indx_2[count]==2:
+                        little_path[count][s]+=brillouin_primitive_vectors_2d[1,s]
+                    else:
+                        little_path[count][s]+=brillouin_primitive_vectors_2d[1,s]+brillouin_primitive_vectors_2d[0,s]
+            count+=1
+        new_little_paths.extend(little_path)
 
-#TESTING INPUT
+    ### reordering the list
+    number_little_paths=int(len(new_little_paths)/number_vertices)
+    new_little_paths=np.reshape(new_little_paths,(number_little_paths,number_vertices,3))
+    ### considering only coordinates in the chosen plane
+    new_little_paths_projected=np.zeros((number_little_paths,number_vertices,2),dtype=float)
+    for i in range(number_little_paths):
+        for j in range(number_vertices):
+            for r in range(2):
+                new_little_paths_projected[i,j,r]+=(new_little_paths[i,j]@brillouin_primitive_vectors_2d[r])
+    fig,ax=plt.subplots()
+    patches=[]
+    for i in range(number_little_paths):
+        polygon=Polygon(new_little_paths_projected[i],closed=True,fill=None,edgecolor='c')
+        patches.append(polygon)
+    p = PatchCollection(patches,cmap=matplotlib.cm.jet,alpha=0.4)
+    colors = 100*np.random.rand(len(patches))
+    p.set_array(np.array(colors))
+    ax.add_collection(p)
+    fig=plt.figure()
+    ax=fig.add_subplot(projection='3d')
+    ax.scatter(k_points_list[:,0],k_points_list[:,1],k_points_list[:,3])
+    plt.show()
+
+### TESTING INPUT
 if __name__ == "__main__":
     import timeit
     import os
@@ -1328,242 +1307,141 @@ if __name__ == "__main__":
     brillouin_primitive_vectors_3d[0]=[1,0,0]
     brillouin_primitive_vectors_3d[1]=[0,1,0]
     brillouin_primitive_vectors_3d[2]=[0,0,1]
-    brillouin_primitive_vectors_2d=np.zeros((2,3),dtype=float)
     chosen_plane=[1,1,0]
-    symmetries=[[0,0,np.pi/2]]
-    initial_grid_spacing=0.1
-    refinment_iteration=0
-    refinment_spacing=0.1
-    threshold=0.001
-    epsilon=0.000000001
-    default_gridding=1000
+    symmetries=[[0,0,0]]
+    grid_spacing=0.1
+    refinment_iterations=3
+    refinment_spacing=0.001
+    threshold_symmetry=0.001
+    threshold_minimal_refinment=0.000000001
+    default_gridding=100
     count=0
     shift_in_plane=[0,0]
     shift_in_space=[0,0,0]
+    covering_BZ=True
 
-    ###LITTLE TRIANGLES
-    refined_k_points_list,triangles=k_points_generator_2D(
-        brillouin_primitive_vectors_3d,
-        initial_grid_spacing,
-        chosen_plane,
-        brillouin_primitive_vectors_2d,
-        True,
-        refinment_spacing,
-        symmetries,
-        threshold,
-        refinment_iteration,
-        shift_in_space,
-        shift_in_space,
-        count
-        )
-    
-    print(brillouin_primitive_vectors_3d)
-    print(brillouin_primitive_vectors_2d)
-    number_triangles=len(triangles)
-    triangle=[0]*3
-    new_triangles=[]
-    for i in range(number_triangles):
-        indx_1=[triangles[i][r][0] for r in range(3)]
-        indx_2=[triangles[i][r][1] for r in range(3)]
-        count=0
-        for indx in indx_1:
-            triangle[count]=[refined_k_points_list[indx,s] for s in range(3)]
-            if indx_2[count]!=0:
-                for s in range(3):
-                    if indx_2[count]==1:
-                        triangle[count][s]+=brillouin_primitive_vectors_2d[0,s]
-                    elif indx_2[count]==2:
-                        triangle[count][s]+=brillouin_primitive_vectors_2d[1,s]
-                    else:
-                        triangle[count][s]+=brillouin_primitive_vectors_2d[1,s]+brillouin_primitive_vectors_2d[0,s]
-            count+=1
-        new_triangles.extend(triangle)
-    print(new_triangles) 
-    new_triangles=np.reshape(new_triangles,(number_triangles,3,3))
-    new_triangles=new_triangles[:,:,:2]
-    fig,ax=plt.subplots()
-    patches=[]
-    
-    for i in range(number_triangles):
-        polygon=Polygon(new_triangles[i], closed=True, fill=None, edgecolor='b')
-        patches.append(polygon)
-    
-    p = PatchCollection(patches,cmap=matplotlib.cm.jet,alpha=0.4)
-    colors = 100*np.random.rand(len(patches))
-    p.set_array(np.array(colors))
-    ax.add_collection(p)
-    plt.show()
-    fig=plt.figure()
-    ax=fig.add_subplot(projection='3d')
-    ax.scatter(refined_k_points_list[:,0],refined_k_points_list[:,1],refined_k_points_list[:,3])
-    plt.show()
-
-    ###LITTLE PARALLELOGRAMS
-    not_refined_k_points_list,parallelograms=k_points_generator_2D(
-        brillouin_primitive_vectors_3d,
-        brillouin_primitive_vectors_3d,
-        initial_grid_spacing,
-        chosen_plane,
-        brillouin_primitive_vectors_2d,
-        default_gridding,
-        False,
-        refinment_spacing,
-        symmetries,
-        threshold,
-        refinment_iteration,
-        shift_in_plane,
-        shift_in_space,
-        count
-        )
-    ###print(not_refined_k_points_list)
-    ##print(parallelograms)
-    number_parallelograms=len(parallelograms)
-    print(np.shape(not_refined_k_points_list))
-    print(np.shape(parallelograms))
-    #print(parallelograms)
-    parallelogram=[0]*4
-    new_parallelograms_tmp=[]
-    for i in range(number_parallelograms):
-        indx_1=[parallelograms[i][r][0] for r in range(4)]
-        indx_2=[parallelograms[i][r][1] for r in range(4)]
-        count=0
-        for indx in indx_1:
-            parallelogram[count]=[not_refined_k_points_list[indx,s] for s in range(3)]
-            if indx_2[count]!=0:
-                for s in range(3):
-                    if indx_2[count]==1:
-                        parallelogram[count][s]+=brillouin_primitive_vectors_2d[0,s]
-                    elif indx_2[count]==2:
-                        parallelogram[count][s]+=brillouin_primitive_vectors_2d[1,s]
-                    else:
-                        parallelogram[count][s]+=brillouin_primitive_vectors_2d[1,s]+brillouin_primitive_vectors_2d[0,s]
-            count+=1
-        new_parallelograms_tmp.extend(parallelogram)
-    print(np.shape(new_parallelograms_tmp))
-    number_parallelograms=int(len(new_parallelograms_tmp)/4)
-    new_parallelograms=np.reshape(new_parallelograms_tmp,(number_parallelograms,4,3))
-    print(new_parallelograms)
-    new_parallelograms=new_parallelograms[:,:,:2]
-    fig,ax=plt.subplots()
-    patches=[]
-    for i in range(number_parallelograms):
-        polygon=Polygon(new_parallelograms[i], closed=True, fill=None, edgecolor='c')
-        patches.append(polygon)
-    
-    print(new_parallelograms)
-    p = PatchCollection(patches,cmap=matplotlib.cm.jet,alpha=0.4)
-    colors = 100*np.random.rand(len(patches))
-    p.set_array(np.array(colors))
-    ax.add_collection(p)
-    plt.show()
-    ##fig=plt.figure()
-    ##ax=fig.add_subplot(projection='3d')
-    ##ax.scatter(not_refined_k_points_list[:,0],not_refined_k_points_list[:,1],not_refined_k_points_list[:,3])
-    ###plt.show()
-
-    position_little_paths_to_refine=[31,32,33]
-    for i in position_little_paths_to_refine:
-        print(parallelograms[i])
-    number_vertices=4,
-    normalized_brillouin_primitive_vectors_2d=np.zeros((2,3),dtype=float)
-    for i in range(2):
-        normalized_brillouin_primitive_vectors_2d[i]=brillouin_primitive_vectors_2d[i]/np.dot(brillouin_primitive_vectors_2d[i],brillouin_primitive_vectors_2d[i])
-    
-    #print("prima", not_refined_k_points_list)
-    #print(parallelograms)
-    #print(position_little_paths_to_refine,
-    #    number_vertices,
-    #    not_refined_k_points_list,
-    #    refinment_iteration,
-    #    symmetries,
-    #    threshold,
-    #    brillouin_primitive_vectors_3d,
-    #    brillouin_primitive_vectors_2d,
-    #    chosen_plane,
-    #    normalized_brillouin_primitive_vectors_2d,
-    #    epsilon)
-    #
-    not_refined_k_points_list,parallelograms=dynamical_refinment_little_paths_2d(
-        parallelograms,
-        position_little_paths_to_refine,
-        number_vertices,
-        not_refined_k_points_list,
-        refinment_iteration,
-        symmetries,
-        threshold,
-        brillouin_primitive_vectors_3d,
-        brillouin_primitive_vectors_2d,
-        chosen_plane,
-        normalized_brillouin_primitive_vectors_2d,
-        epsilon
-    )  
-    #print("dopo", not_refined_k_points_list)
-    #print(parallelograms)
-    number_parallelograms=len(parallelograms)
-    #print(np.shape(not_refined_k_points_list))
-    #print(np.shape(parallelograms))
-    #print(parallelograms)
-    parallelogram=[0]*4
-    new_parallelograms_tmp=[]
-    for i in range(number_parallelograms):
-        indx_1=[parallelograms[i][r][0] for r in range(4)]
-        indx_2=[parallelograms[i][r][1] for r in range(4)]
-        count=0
-        for indx in indx_1:
-            parallelogram[count]=[not_refined_k_points_list[indx,s] for s in range(3)]
-            if indx_2[count]!=0:
-                for s in range(3):
-                    if indx_2[count]==1:
-                        parallelogram[count][s]+=brillouin_primitive_vectors_2d[0,s]
-                    elif indx_2[count]==2:
-                        parallelogram[count][s]+=brillouin_primitive_vectors_2d[1,s]
-                    else:
-                        parallelogram[count][s]+=brillouin_primitive_vectors_2d[1,s]+brillouin_primitive_vectors_2d[0,s]
-            count+=1
-        new_parallelograms_tmp.extend(parallelogram)
-    print(np.shape(new_parallelograms_tmp))
-    number_parallelograms=int(len(new_parallelograms_tmp)/4)
-    new_parallelograms=np.reshape(new_parallelograms_tmp,(number_parallelograms,4,3))
-    #print(new_parallelograms)
-    new_parallelograms=new_parallelograms[:,:,:2]
-    fig,ax=plt.subplots()
-    patches=[]
-    for i in range(number_parallelograms):
-        polygon=Polygon(new_parallelograms[i], closed=True, fill=None, edgecolor='c')
-        patches.append(polygon)
-    
-    #print(new_parallelograms)
-    p = PatchCollection(patches,cmap=matplotlib.cm.jet,alpha=0.4)
-    colors = 100*np.random.rand(len(patches))
-    p.set_array(np.array(colors))
-    ax.add_collection(p)
-    plt.show()
-    ##fig=plt.figure()
-    ##ax=fig.add_subplot(projection='3d')
-    ##ax.scatter(not_refined_k_points_list[:,0],not_refined_k_points_list[:,1],not_refined_k_points_list[:,3])
-    ##plt.show()
-
-    ##dynamical refinment (parallelograms)
+    ### LITTLE PARALLELOGRAMS
+    ##number_vertices=4
+    ##not_refined_k_points_list,parallelograms=k_points_generator_2D(
+    ##    brillouin_primitive_vectors_3d,
+    ##    brillouin_primitive_vectors_3d,
+    ##    chosen_plane,
+    ##    grid_spacing,
+    ##    covering_BZ,
+    ##    default_gridding,
+    ##    False,
+    ##    refinment_spacing,
+    ##    refinment_iterations,
+    ##    symmetries,
+    ##    threshold_symmetry,
+    ##    shift_in_plane,
+    ##    shift_in_space,
+    ##    0
+    ##)
     ##print(not_refined_k_points_list)
     ##print(parallelograms)
-    ##indices=[0,1,2,3,4,4,4,5]
-    ##subset_parallelograms=parallelograms[indices]
-    ##print(subset_parallelograms)
-    ##normalized_brillouin_primitive_vectors_2d=np.zeros((2,3))
-    ##for i in range(2):
-    ##    for r in range(3):
-    ##        normalized_brillouin_primitive_vectors_2d[i,r]=brillouin_primitive_vectors_2d[i,r]/np.dot(brillouin_primitive_vectors_2d[i],brillouin_primitive_vectors_2d[i])
-    ##
-    ##not_refined_k_points_list, parallelograms=dynamical_refinment_little_paths_2d(
-    ##    parallelograms,
-    ##    indices,
-    ##    4,
-    ##    not_refined_k_points_list,
-    ##    refinment_iteration,
-    ##    symmetries,
-    ##    threshold,
+    ##printing_covering_BZ_2D(
     ##    brillouin_primitive_vectors_3d,
-    ##    brillouin_primitive_vectors_2d,
     ##    chosen_plane,
-    ##    normalized_brillouin_primitive_vectors_2d
+    ##    not_refined_k_points_list,
+    ##    parallelograms,
+    ##    number_vertices
     ##)
+    ###DYNAMICAL REFINMENT
+    ##brillouin_primitive_vectors_2d=np.zeros((2,3),dtype=float)
+    ##normalized_brillouin_primitive_vectors_2d=np.zeros((2,3),dtype=float)
+    ##count = 0
+    ##for i in range(3):
+    ##    if chosen_plane[i]!=0:
+    ##        brillouin_primitive_vectors_2d[count] = brillouin_primitive_vectors_3d[i]
+    ##        normalized_brillouin_primitive_vectors_2d[count] = brillouin_primitive_vectors_2d[count]/(brillouin_primitive_vectors_2d[count]@brillouin_primitive_vectors_2d[count])
+    ##        count += 1
+    ##
+    ##position_little_paths_to_refine=[2,58,68]
+    ##not_refined_k_points_list,parallelograms=dynamical_refinment_little_paths_2D(
+    ##    parallelograms,
+    ##    not_refined_k_points_list,
+    ##    position_little_paths_to_refine,
+    ##    number_vertices,
+    ##    refinment_iterations,
+    ##    symmetries,
+    ##    threshold_symmetry,
+    ##    brillouin_primitive_vectors_3d,
+    ##    chosen_plane,
+    ##    brillouin_primitive_vectors_2d,
+    ##    normalized_brillouin_primitive_vectors_2d,
+    ##    threshold_minimal_refinment,
+    ##    default_gridding
+    ##)
+    ##printing_covering_BZ_2D(
+    ##    brillouin_primitive_vectors_3d,
+    ##    chosen_plane,
+    ##    not_refined_k_points_list,
+    ##    parallelograms,
+    ##    number_vertices
+    ##)
+
+    ###LITTLE TRIANGLES
+    number_vertices=3
+    refined_k_points_list,triangles=k_points_generator_2D(
+        brillouin_primitive_vectors_3d,
+        brillouin_primitive_vectors_3d,
+        chosen_plane,
+        grid_spacing,
+        covering_BZ,
+        default_gridding,
+        True,
+        refinment_spacing,
+        refinment_iterations,
+        symmetries,
+        threshold_symmetry,
+        shift_in_plane,
+        shift_in_space,
+        0
+        )
+    ##print(refined_k_points_list)
+    ##print(triangles)
+    printing_covering_BZ_2D(
+        brillouin_primitive_vectors_3d,
+        chosen_plane,
+        refined_k_points_list,
+        triangles,
+        number_vertices
+    )
+    ###DYNAMICAL REFINMENT
+    brillouin_primitive_vectors_2d=np.zeros((2,3),dtype=float)
+    normalized_brillouin_primitive_vectors_2d=np.zeros((2,3),dtype=float)
+    count = 0
+    for i in range(3):
+        if chosen_plane[i]!=0:
+            brillouin_primitive_vectors_2d[count] = brillouin_primitive_vectors_3d[i]
+            normalized_brillouin_primitive_vectors_2d[count] = brillouin_primitive_vectors_2d[count]/(brillouin_primitive_vectors_2d[count]@brillouin_primitive_vectors_2d[count])
+            count += 1
+    position_little_paths_to_refine=[37]
+    not_refined_k_points_list,triangles_plus=dynamical_refinment_little_paths_2D(
+        triangles,
+        refined_k_points_list,
+        position_little_paths_to_refine,
+        number_vertices,
+        refinment_iterations,
+        symmetries,
+        threshold_symmetry,
+        brillouin_primitive_vectors_3d,
+        chosen_plane,
+        brillouin_primitive_vectors_2d,
+        normalized_brillouin_primitive_vectors_2d,
+        threshold_minimal_refinment,
+        default_gridding
+    )
+    print("dopo",triangles_plus)
+    all_triangles=[]
+    all_triangles.extend(triangles)
+    all_triangles.extend(triangles_plus)
+    all_triangles=np.reshape(all_triangles,(int(len(all_triangles)),3,2))
+    printing_covering_BZ_2D(
+        brillouin_primitive_vectors_3d,
+        chosen_plane,
+        not_refined_k_points_list,
+        all_triangles,
+        number_vertices
+    )
